+++ dcpuhw.c
@@ -30,6 +30,14 @@ static int HWM_FreeAll(struct isiInfo *info)
 	return 0;
 }
 
+static int HWM_QAttach(struct isiInfo *info, int32_t point, struct isiInfo *dev, int32_t devpoint)
+{
+	if(!dev) return -1;
+	if(point == ISIAT_UP) return -1;
+	if(dev->id.objtype == ISIT_MEM6416) return 0;
+	if(info->busdev.count < 1 && dev->id.objtype != ISIT_DCPU) return ISIERR_MISSPREREQ;
+	return 0;
+}
 static int HWM_DeviceAdd(struct isiInfo *info, int32_t point, struct isiInfo *dev, int32_t devpoint)
 {
 	if(!dev) return -1;
@@ -40,15 +48,7 @@ static int HWM_DeviceAdd(struct isiInfo *info, int32_t point, struct isiInfo *de
 
 static int HWM_Attached(struct isiInfo *info, int32_t point, struct isiInfo *dev, int32_t devpoint)
 {
-	size_t k;
-	size_t hs;
 	isilog(L_DEBUG, "hwm: updating attachments c=%d\n", info->busdev.count);
-	hs = info->busdev.count;
-	for(k = 0; k < hs; k++) {
-		if(info->busdev.table[k].t) {
-			info->busdev.table[k].t->mem = info->mem;
-		}
-	}
 	return 0;
 }
 
@@ -75,10 +75,11 @@ static int HWM_Query(struct isiInfo *info, struct isiInfo *src, int32_t lsindex,
 		msg[1] = (uint16_t)(hs - 1);
 		isilog(L_DEBUG, "hwm-reset-all c=%ld\n", hs);
 		for(h = 1; h < hs; h++) {
-			dev = info->busdev.table[h].t;
-			if(dev->c->Reset) dev->c->Reset(dev);
-			if(!isi_message_dev(info, h, msg, len, mtime)) {
-				isilog(L_DEBUG, "hwm-reset: %s %ld\n", dev->meta->name, h);
+			if(!isi_getindex_dev(info, h, &dev)) {
+				if(dev->c->Reset) dev->c->Reset(dev);
+				if(!isi_message_dev(info, h, msg, len, mtime)) {
+					isilog(L_DEBUG, "hwm-reset: %s %ld\n", dev->meta->name, h);
+				}
 			}
 		}
 		return 0;
@@ -86,19 +87,21 @@ static int HWM_Query(struct isiInfo *info, struct isiInfo *src, int32_t lsindex,
 	case ISE_XINT:
 		if(h >= hs) {
 			isilog(L_DEBUG, "hwm: %ld out of range.\n", h);
+			r = -1;
 			break;
 		}
-		dev = info->busdev.table[h].t;
-		if(msg[0] == ISE_QINT && dev->meta->meta) {
-			struct isidcpudev *mid = (struct isidcpudev *)dev->meta->meta;
-			msg[2] = (uint16_t)(mid->devid);
-			msg[3] = (uint16_t)(mid->devid >> 16);
-			msg[4] = mid->verid;
-			msg[5] = (uint16_t)(mid->mfgid);
-			msg[6] = (uint16_t)(mid->mfgid >> 16);
-			r = 0;
-		}
-		r = isi_message_dev(info, h, msg, len, mtime);
+		if(!isi_getindex_dev(info, h, &dev)) {
+			if(msg[0] == ISE_QINT && dev->meta->meta) {
+				struct isidcpudev *mid = (struct isidcpudev *)dev->meta->meta;
+				msg[2] = (uint16_t)(mid->devid);
+				msg[3] = (uint16_t)(mid->devid >> 16);
+				msg[4] = mid->verid;
+				msg[5] = (uint16_t)(mid->mfgid);
+				msg[6] = (uint16_t)(mid->mfgid >> 16);
+				r = 0;
+			}
+			r = isi_message_dev(info, h, msg, len, mtime);
+		} else r = -1;
 		return r;
 	default:
 		break;
@@ -113,9 +116,10 @@ static int HWM_Run(struct isiInfo *info, struct timespec crun)
 	struct isiInfo *dev;
 	hs = info->busdev.count;
 	for(k = 1; k < hs; k++) {
-		dev = info->busdev.table[k].t;
-		if(dev->c->RunCycles) {
-			dev->c->RunCycles(dev, crun);
+		if(!isi_getindex_dev(info, k, &dev)) {
+			if(dev->c->RunCycles) {
+				dev->c->RunCycles(dev, crun);
+			}
 		}
 	}
 	return 0;
@@ -124,6 +128,7 @@ static int HWM_Run(struct isiInfo *info, struct timespec crun)
 static struct isiInfoCalls HWMCalls = {
 	.RunCycles = HWM_Run,
 	.MsgIn = HWM_Query,
+	.QueryAttach = HWM_QAttach,
 	.Attach = HWM_DeviceAdd,
 	.Attached = HWM_Attached,
 	.Delete = HWM_FreeAll