+++ hardware/floppy_m35.c
@@ -9,6 +9,13 @@
 #define ERR_M35_SECTOR 5
 #define ERR_M35_BROKEN 0xffff
 
+#define ST_M35_NO_MEDIA 0
+#define ST_M35_IDLE 1
+#define ST_M35_IDLEP 2
+#define ST_M35_BUSY 3
+#define ST_M35_BUSYREAD 3
+#define ST_M35_BUSYWRITE 4
+
 /* state while this device exists/active (runtime volatile) */
 struct Disk_M35FD_rvstate {
 	uint16_t track;
@@ -57,7 +64,7 @@ void Disk_M35FD_Register()
 
 static int Disk_M35FD_QAttach(struct isiInfo *info, int32_t point, struct isiInfo *dev, int32_t devpoint)
 {
-	if(!dev || dev->id.objtype != ISIT_DISK) return -1;
+	if(!dev || (point != ISIAT_UP && dev->id.objtype != ISIT_DISK)) return -1;
 	return 0;
 }
 
@@ -99,10 +106,10 @@ static int Disk_M35FD_HWI(struct isiInfo *info, struct isiInfo *src, uint16_t *m
 	switch(msg[0]) {
 	case 0:
 		if(media) {
-			if(dev->oper) {
-				msg[1] = 3;
+			if(dev->oper >= ST_M35_BUSY) {
+				msg[1] = ST_M35_BUSY;
 			} else {
-				msg[1] = 1; // TODO RW vs RO floppy
+				msg[1] = dev->oper;
 			}
 		} else {
 			msg[1] = 0;
@@ -117,7 +124,7 @@ static int Disk_M35FD_HWI(struct isiInfo *info, struct isiInfo *src, uint16_t *m
 	case 2:
 		msg[1] = 0;
 		lerr = dev->errcode;
-		if(dev->oper != 0) dev->errcode = ERR_M35_BUSY;
+		if(dev->oper >= ST_M35_BUSY) dev->errcode = ERR_M35_BUSY;
 		else if(!media) dev->errcode = ERR_M35_NO_MEDIA;
 		else if(msg[3] >= 1440) dev->errcode = ERR_M35_SECTOR;
 		else {
@@ -128,7 +135,7 @@ static int Disk_M35FD_HWI(struct isiInfo *info, struct isiInfo *src, uint16_t *m
 			dseek.block = dev->seeksector >> 2;
 			isi_message_dev(info, 0, (uint16_t*)&dseek, 4, crun);
 			dev->rwaddr = msg[4];
-			dev->oper = 1;
+			dev->oper = ST_M35_BUSYREAD;
 			msg[1] = 1;
 			Disk_M35FD_statechange(info, &crun);
 		}
@@ -137,10 +144,10 @@ static int Disk_M35FD_HWI(struct isiInfo *info, struct isiInfo *src, uint16_t *m
 	case 3:
 		msg[1] = 0;
 		lerr = dev->errcode;
-		if(dev->oper != 0) dev->errcode = ERR_M35_BUSY;
+		if(dev->oper >= ST_M35_BUSY) dev->errcode = ERR_M35_BUSY;
 		else if(!media) dev->errcode = ERR_M35_NO_MEDIA;
 		else if(msg[3] >= 1440) dev->errcode = ERR_M35_SECTOR;
-		else if(0) dev->errcode = ERR_M35_PROTECT; /* TODO write_protect */
+		else if(isi_disk_isreadonly(media)) dev->errcode = ERR_M35_PROTECT; /* TODO write_protect */
 		else {
 			dev->seeksector = msg[3];
 			dev->seektrack = dev->seeksector / 18;
@@ -153,7 +160,7 @@ static int Disk_M35FD_HWI(struct isiInfo *info, struct isiInfo *src, uint16_t *m
 				dss->svbuf[i] = isi_hw_rdmem((isiram16)info->mem, mc);
 				mc++;
 			}
-			dev->oper = 2;
+			dev->oper = ST_M35_BUSYWRITE;
 			r = 512;
 			msg[1] = 1;
 			Disk_M35FD_statechange(info, &crun);
@@ -169,11 +176,15 @@ static int Disk_M35FD_Tick(struct isiInfo *info, struct timespec crun)
 	struct Disk_M35FD_rvstate *dev = (struct Disk_M35FD_rvstate*)info->rvstate;
 	struct Disk_M35FD_svstate *dss = (struct Disk_M35FD_svstate*)info->svstate;
 	struct isiInfo *media = 0;
+	if(!info->nrun.tv_sec) {
+		info->nrun.tv_sec = crun.tv_sec;
+		info->nrun.tv_nsec = crun.tv_nsec;
+	}
 	while(isi_time_lt(&info->nrun, &crun)) {
-		if(dev->oper) {
+		if(dev->oper >= ST_M35_BUSY) {
 			if(isi_getindex_dev(info, 0, &media) || !media->c->MsgIn) {
 				dev->errcode = ERR_M35_EJECT;
-				dev->oper = 0;
+				dev->oper = ST_M35_NO_MEDIA;
 				Disk_M35FD_statechange(info, &info->nrun);
 				continue;
 			}
@@ -193,26 +204,33 @@ static int Disk_M35FD_Tick(struct isiInfo *info, struct timespec crun)
 					uint16_t *dr;
 					isi_disk_getblock(media, (void**)&dr);
 					dr += (512 * (dev->seeksector & 3));
-					if(dev->oper == 1) {
+					if(dev->oper == ST_M35_BUSYREAD) {
 						uint16_t mc;
 						mc = dev->rwaddr;
 						for(int i = 0; i < 512; i++) {
 							isi_hw_wrmem((isiram16)info->mem, mc, dr[i]);
 							mc++;
 						}
-					} else if(dev->oper == 2) {
+					} else if(dev->oper == ST_M35_BUSYWRITE) {
 						for(int i = 0; i < 512; i++) {
 							dr[i] = dss->svbuf[i];
 						}
 					}
-					dev->oper = 0;
+					dev->oper = isi_disk_isreadonly(media)? ST_M35_IDLEP:ST_M35_IDLE;
 					Disk_M35FD_statechange(info, &info->nrun);
 				}
 			}
 		} else {
-			if(!info->nrun.tv_sec) {
-				info->nrun.tv_sec = crun.tv_sec;
-				info->nrun.tv_nsec = crun.tv_nsec;
+			if(isi_getindex_dev(info, 0, &media)) {
+				if(dev->oper != ST_M35_NO_MEDIA) {
+					dev->oper = ST_M35_NO_MEDIA;
+					Disk_M35FD_statechange(info, &info->nrun);
+				}
+			} else {
+				if(dev->oper == ST_M35_NO_MEDIA) {
+					dev->oper = isi_disk_isreadonly(media)? ST_M35_IDLEP:ST_M35_IDLE;
+					Disk_M35FD_statechange(info, &info->nrun);
+				}
 			}
 			isi_addtime(&info->nrun, 50000000);
 			dev->sector += 1535;