+++ hwtbl.c
@@ -4,6 +4,7 @@
 #include "dcpuhw.h"
 
 /* Hardware functions */
+void cemei_register();
 void DCPU_Register();
 void Memory_Register();
 void Disk_Register();
@@ -19,9 +20,10 @@ void KaiHIC_Register();
 
 void isi_register_objects()
 {
-	DCPU_Register();
 	Memory_Register();
+	cemei_register();
 	Disk_Register();
+	DCPU_Register();
 	DCPUBUS_Register();
 	Keyboard_Register();
 	Clock_Register();